+++ modules/express-route/r-logs.tf
@@ -1,6 +1,6 @@
 module "express_route_circuit_diagnostic_settings" {
   source  = "claranet/diagnostic-settings/azurerm"
-  version = "6.2.0"
+  version = "~> 6.3.0"
 
   resource_id           = azurerm_express_route_circuit.erc.id
   logs_destinations_ids = var.logs_destinations_ids