+++ r-express-route.tf
@@ -14,7 +14,7 @@ module "express_route" {
   name_slug   = var.name_slug
   name_suffix = var.name_suffix
 
-  logs_destinations_ids   = var.logs_destinations_ids
+  logs_destinations_ids   = coalesce(var.express_route_logs_destinations_ids, var.logs_destinations_ids)
   logs_categories         = var.express_route_logs_categories
   logs_metrics_categories = var.express_route_logs_metrics_categories
   logs_retention_days     = var.logs_retention_days