+++ variables-logs.tf
@@ -7,6 +7,7 @@ List of destination resources IDs for logs diagnostic destination.
 Can be `Storage Account`, `Log Analytics Workspace` and `Event Hub`. No more than one of each can be set.
 If you want to specify an Azure EventHub to send logs and metrics to, you need to provide a formated string with both the EventHub Namespace authorization send ID and the EventHub name (name of the queue to use in the Namespace) separated by the `|` character.
 EOD
+  nullable    = false
 }
 
 variable "logs_retention_days" {
@@ -34,6 +35,16 @@ variable "express_route_custom_diagnostic_settings_name" {
   default     = "default"
 }
 
+variable "express_route_logs_destinations_ids" {
+  type        = list(string)
+  description = <<EOD
+List of destination resources IDs for logs diagnostic destination for `Express Route` resource.
+Can be `Storage Account`, `Log Analytics Workspace` and `Event Hub`. No more than one of each can be set.
+If you want to specify an Azure EventHub to send logs and metrics to, you need to provide a formated string with both the EventHub Namespace authorization send ID and the EventHub name (name of the queue to use in the Namespace) separated by the `|` character.
+EOD
+  default     = null
+}
+
 variable "firewall_logs_categories" {
   type        = list(string)
   description = "Log categories to send to destinations."
@@ -52,6 +63,16 @@ variable "firewall_custom_diagnostic_settings_name" {
   default     = "default"
 }
 
+variable "firewall_logs_destinations_ids" {
+  type        = list(string)
+  description = <<EOD
+List of destination resources IDs for logs diagnostic destination for `Azure Firewall` resource.
+Can be `Storage Account`, `Log Analytics Workspace` and `Event Hub`. No more than one of each can be set.
+If you want to specify an Azure EventHub to send logs and metrics to, you need to provide a formated string with both the EventHub Namespace authorization send ID and the EventHub name (name of the queue to use in the Namespace) separated by the `|` character.
+EOD
+  default     = null
+}
+
 variable "vpn_gateway_logs_categories" {
   type        = list(string)
   description = "Log categories to send to destinations."
@@ -68,4 +89,14 @@ variable "vpn_gateway_custom_diagnostic_settings_name" {
   description = "Custom name of the diagnostics settings, name will be 'default' if not set."
   type        = string
   default     = "default"
+}
+
+variable "vpn_gateway_logs_destinations_ids" {
+  type        = list(string)
+  description = <<EOD
+List of destination resources IDs for logs diagnostic destination for `VPN Gateway` resource.
+Can be `Storage Account`, `Log Analytics Workspace` and `Event Hub`. No more than one of each can be set.
+If you want to specify an Azure EventHub to send logs and metrics to, you need to provide a formated string with both the EventHub Namespace authorization send ID and the EventHub name (name of the queue to use in the Namespace) separated by the `|` character.
+EOD
+  default     = null
 }